Refactoring a bit
